+++ demos/employees/src/main/java/module-info.java
@@ -29,6 +29,7 @@ module is.codion.demos.employees {
 	requires is.codion.framework.server;
 	requires is.codion.plugin.jasperreports;
 	requires is.codion.plugin.flatlaf.intellij.themes;
+	requires is.codion.plugin.swing.robot;
 
 	exports is.codion.demos.employees.domain
 					to is.codion.framework.domain, is.codion.framework.db.local;